# static/swagger.json
{
    "swagger": "2.0",
    "basePath": "/api",
    "paths": {
...
        "/organize/user/{organizer_id}": {
            "parameters": [
                {
                    "in": "path",
                    "description": "Firebase User ID of the organizer",
                    "name": "organizer_id",
                    "required": true,
                    "type": "string"
                }
            ],
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Assign a trip to an organizer",
                "operationId": "Assign a trip to an organizer",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Trip%20ID"
                        }
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "organize"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Complete%20Trip"
                            }
                        }
                    }
                },
                "summary": "Get a list of all trips organized by the specified organizer",
                "operationId": "Get a list of all trips organized by the specified organizer",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "organize"
                ]
            },
            "delete": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Remove a trip from an organizer",
                "operationId": "Remove a trip from an organizer",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Trip%20ID"
                        }
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "organize"
                ]
            }
        },
...
        "/registration/user/self": {
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Register the current user for a trip",
                "operationId": "Register the current user for a trip",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Trip%20ID"
                        }
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "registration"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Complete%20User%20Trip"
                            }
                        }
                    }
                },
                "summary": "Get a list of all trips the current user is registered for",
                "operationId": "Get a list of all trips the current user is registered for",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "registration"
                ]
            },
            "delete": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Cancel the current user's registration for a trip",
                "operationId": "Cancel the current user's registration for a trip",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Trip%20ID"
                        }
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "registration"
                ]
            },
            "put": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Modify the current user's registration for a trip",
                "operationId": "Modify the current user's registration for a trip",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Registration%20without%20ID"
                        }
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "registration"
                ]
            }
        },
        "/registration/user/{firebase_id}": {
            "parameters": [
                {
                    "in": "path",
                    "description": "Firebase User ID",
                    "name": "firebase_id",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Complete%20User%20Trip"
                            }
                        }
                    }
                },
                "summary": "Get a list of all trips a specific user is registered for",
                "operationId": "Get a list of all trips a specific user is registered for",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "registration"
                ]
            }
        },
...
        "/user/": {
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Add a new user to the database",
                "operationId": "Create a new user",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Create%20User"
                        }
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Complete%20User"
                            }
                        }
                    }
                },
                "summary": "Get a list of all users in the database",
                "operationId": "List all registered users",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        },
        "/user/friends/self": {
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Add a friend for the current user",
                "operationId": "Add a friend for the current user",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/User%20ID"
                        }
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Complete%20User"
                            }
                        }
                    }
                },
                "summary": "Get a list of the current user's friends",
                "operationId": "Get a list of the current user's friends",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            },
            "delete": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Remove a friend for the current user",
                "operationId": "Remove a friend for the current user",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/User%20ID"
                        }
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        },
        "/user/friends/{firebase_id}": {
            "parameters": [
                {
                    "in": "path",
                    "description": "Firebase User ID",
                    "name": "firebase_id",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Complete%20User"
                            }
                        }
                    }
                },
                "summary": "Get a list of friends for the specified user",
                "operationId": "Get a list of friends for the specified user",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        },
        "/user/organizer": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Complete%20User"
                            }
                        }
                    }
                },
                "summary": "Get a list of all registered organizers",
                "operationId": "List all registered organizers",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        },
        "/user/self": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Complete%20User"
                        }
                    }
                },
                "summary": "Get the current user's data",
                "operationId": "Get the current user's data",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            },
            "delete": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Delete the current user",
                "operationId": "Delete the current user",
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            },
            "put": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Update the current user's data",
                "operationId": "Update the current user's data",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/User%20Self"
                        }
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        },
        "/user/suggestedFriends": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Complete%20User"
                            }
                        }
                    }
                },
                "summary": "Get a list of suggested friends for the current user",
                "operationId": "Get a list of suggested friends for the current user",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        },
        "/user/{firebase_id}": {
            "parameters": [
                {
                    "in": "path",
                    "description": "Firebase User ID",
                    "name": "firebase_id",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Complete%20User"
                        }
                    }
                },
                "summary": "Get the specified user",
                "operationId": "Get a specific user",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        }
    },
    "info": {
        "title": "MountainMates API",
        "version": "0.1",
        "description": "This API facilitates hiking trips, participant registration, and trip organization."
    },
    "produces": [
        "application/json"
    ],
    "consumes": [
        "application/json"
    ],
    "securityDefinitions": {
        "Bearer Auth": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    },
...
}